refactor: move fillBufferWithRandomBytes to utils
